requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Redis==0.4.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
bcrypt==4.0.1
redis==4.6.0
alembic==1.12.0
werkzeug==2.3.7
shapely==2.0.2
bleach==6.1.0
passlib==1.7.4
